pyproject.toml
[project]
name = "auto-refactor-mcp-server"
version = "0.1.0"
description = "Auto-Refactor MCP Server - Intelligent automated code refactoring"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Arun Kumar", email = "your.email@example.com" }
]
keywords = ["mcp", "refactoring", "code-analysis", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "mcp>=1.2.0",
    "fastmcp>=0.1.0",
    "pydantic>=2.0.0",
    "gitpython>=3.1.0",
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "tree-sitter>=0.21.0",
    "tree-sitter-python>=0.21.0",
    "tree-sitter-javascript>=0.21.0",
    "radon>=6.0.0",
    "pytest>=8.0.0",
    "httpx>=0.24.0",
    "astroid>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
